refactored GetCharWeights

+++ Ascii/FreeType.h
@@ -30,6 +30,6 @@ namespace Ascii
 	namespace Cpp
 	{
 		size_t CountBits(const FT_Bitmap& bitmap);
-		std::map<double, char> GetCharWeights(const std::string& family, const std::string& style, int height, const std::string& charset);
+		std::map<int, char> GetCharWeights(const std::string& family, const std::string& style, int height, const std::string& charset);
 	};
 }
